modularized js code
