+++ public/javascript/utils.js
@@ -1,0 +1,27 @@
+/**
+ * This function is used to create a new element with the given type, attributes and children.
+ *
+ * @param {string} type - The type of the element to create
+ * @param {object} attrs - An object containing the attributes to set on the created element
+ * @param {...HTMLElement|string} children - The children to append to the created element
+ * @returns {HTMLElement}
+ */
+export function elt(type, attrs, ...children) {
+    let node = document.createElement(type);
+
+    // Loop through the attributes object and set attributes on the created element
+    Object.keys(attrs).forEach(key => {
+        node.setAttribute(key, attrs[key]);
+    });
+
+    // Loop through the children array and append them to the created element
+    for (let child of children) {
+        if (typeof child != "string") {
+            node.appendChild(child);
+        } else {
+            node.appendChild(document.createTextNode(child));
+        }
+    }
+
+    return node;
+}
